make carusel in contact page and also working on cards
